Use whole-number quantities only for check-in and check-out.
Quantity inputs now accept integers (step 1). API rejects decimals and
displays strip trailing fractional zeros across dashboard and history.

+++ api/app/schemas.py
@@ -130,13 +130,13 @@ class InventoryItemRead(BaseModel):
 
 class CheckInRequest(BaseModel):
     product_id: UUID
-    quantity: Decimal = Field(gt=0)
+    quantity: int = Field(gt=0)
     note: str | None = None
 
 
 class CheckOutRequest(BaseModel):
     product_id: UUID
-    quantity: Decimal = Field(gt=0)
+    quantity: int = Field(gt=0)
     note: str | None = None
